libsmack: Terminate attribute string
The smack_new_label_from_path function reads an xattr
that may not be null byte terminated. This occurs in the
SMACK64TRANSMUTE case. Technically, the transmute attribute
isn't a label, so this function shouldn't be used to fetch
the value, but we'll let that go. This is just good string
hygiene in any case.

This is an issue because chsmack prints transmute="TRUE0"
without this fix.

Change-Id: I5ed107d5edfe119f3e49f4c16efbdd37927e5b20
Signed-off-by: Casey Schaufler <casey.schaufler@intel.com>
Signed-off-by: Rafal Krypa <r.krypa@samsung.com>